refactor(glass-core): split session.rs into session/ submodules (Phase A)
Convert the 3,785-line session.rs god-file into a session/ directory. The
Glass type, its state (ActiveSession/Backend/PlatformFactory), the a11y-tree
helper free fns, and core helpers stay in mod.rs; each group of operations
moves to its own submodule with an `impl Glass` block:
lifecycle/capture/input/clipboard/window/a11y/wait/baseline.

Code moved verbatim (no logic change). `capture` and `pointer_inner` widen
from private to `pub(super)` since they're now called from sibling submodules.
All public paths preserved via re-exports (glass_core::Glass, Wait*, Scroll*).

The test module stays intact in mod.rs for this checkpoint; Phase B
distributes it beside the code it exercises.

Verified: cargo build/clippy --all-targets -D warnings/test --workspace/fmt --check all green.
