+++ crates/glass-core/src/session/mod.rs
@@ -1,3 +1,6 @@
+//! The `Glass` session manager. The type, its state, and shared helpers live here;
+//! its operations are grouped into submodules (each adds an `impl Glass` block).
+
 use crate::accessibility::{
     element_match, Accessibility, AxContext, AxNode, AxNodeId, AxRole, AxTarget, AxTree,
     ElementCondition, ElementInfo, ElementMatch,
@@ -14,201 +17,20 @@ use crate::platform::{
 };
 use crate::stability::StabilityTracker;
 
-/// Parameters for [`Glass::wait_stable`].
-#[derive(Clone, Debug)]
-pub struct WaitStableParams {
-    pub interval_ms: u64,
-    pub settle_frames: u32,
-    pub tolerance: u8,
-    pub timeout_ms: u64,
-    /// When set, the settle decision compares only this sub-rectangle of each
-    /// frame; the returned frame is still the full window.
-    pub stability_region: Option<Region>,
-    /// When set, watch this window's own region instead of the active window's —
-    /// without changing which window is active.
-    pub window: Option<WindowId>,
-}
+mod a11y;
+mod baseline;
+mod capture;
+mod clipboard;
+mod input;
+mod lifecycle;
+mod wait;
+mod window;
 
-/// Outcome of a wait-until-stable: the final frame and whether it settled
-/// before the timeout.
-#[derive(Clone, Debug)]
-pub struct WaitStableOutcome {
-    pub frame: Frame,
-    pub settled: bool,
-    /// Whether any frame-to-frame change was seen while watching. `settled:true` with
-    /// `saw_motion:false` over a short `observed_ms` is a *brief* quiet window — a slow
-    /// animation can still hide under it, so use `wait_for_region {until:"changes"}` to
-    /// positively assert motion. `settled:true` with `saw_motion:true` means it was moving
-    /// and then quieted.
-    pub saw_motion: bool,
-    /// How long (ms) frames were observed before settling or timing out.
-    pub observed_ms: u64,
-}
-
-/// Parameters for [`Glass::wait_for_element`].
-#[derive(Clone, Debug)]
-pub struct WaitElementParams {
-    pub name: Option<String>,
-    pub role: Option<AxRole>,
-    pub value_contains: Option<String>,
-    pub condition: ElementCondition,
-    pub interval_ms: u64,
-    pub timeout_ms: u64,
-}
-
-/// Outcome of [`Glass::wait_for_element`].
-#[derive(Clone, Debug)]
-pub struct WaitElementOutcome {
-    pub matched: bool,
-    /// The matched element (absent on timeout, and for a satisfied `disappears`).
-    pub element: Option<ElementInfo>,
-    /// Wall-clock milliseconds elapsed when the wait returned.
-    pub elapsed_ms: u64,
-}
-
-/// Wheel notches per scroll step; chosen so a step realizes at most a few rows
-/// (won't skip a virtualized row's realized band). Overridable per call.
-pub const SCROLL_TO_DEFAULT_STEP: u32 = 3;
-/// Overall wall-clock bound for a `scroll_to_element` sweep.
-pub const SCROLL_TO_DEFAULT_TIMEOUT_MS: u64 = 20_000;
-/// Hard cap on scroll steps issued across a full bidirectional sweep, independent
-/// of `timeout_ms` — bounds the sweep even if the caller passes an enormous timeout.
-const SCROLL_TO_MAX_STEPS: u32 = 500;
-/// Milliseconds to let scrolled rows realize in the a11y tree before re-reading.
-/// 250ms is the validated floor on the headless a11y bus: the tree is read once
-/// per step (for both the match and the end-of-scroll comparison), so a settle
-/// shorter than the toolkit's realize latency would read an unchanged tree and
-/// misfire as premature saturation.
-const SCROLL_TO_SETTLE_MS: u64 = 250;
-
-/// A vertical scroll sweep direction.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum ScrollDirection {
-    Down,
-    Up,
-}
-
-impl ScrollDirection {
-    /// The other sweep direction.
-    pub fn opposite(self) -> ScrollDirection {
-        match self {
-            ScrollDirection::Down => ScrollDirection::Up,
-            ScrollDirection::Up => ScrollDirection::Down,
-        }
-    }
-    /// Signed vertical wheel delta (notches): `Down` is positive (wheel-down),
-    /// `Up` negative. Saturates a huge `step` to `i32::MAX` so an absurd caller
-    /// value can't overflow (a plain `step as i32` would wrap, and `-(i32::MIN)`
-    /// panics in debug) — real steps are single digits.
-    pub fn dy(self, step: u32) -> i32 {
-        let s = i32::try_from(step).unwrap_or(i32::MAX);
-        match self {
-            ScrollDirection::Down => s,
-            ScrollDirection::Up => -s,
-        }
-    }
-    /// Parse from a tool string (case-insensitive). `None` for unknown.
-    pub fn from_name(s: &str) -> Option<ScrollDirection> {
-        match s.to_ascii_lowercase().as_str() {
-            "down" => Some(ScrollDirection::Down),
-            "up" => Some(ScrollDirection::Up),
-            _ => None,
-        }
-    }
-}
-
-/// Parameters for [`Glass::scroll_to_element`].
-#[derive(Clone, Debug)]
-pub struct ScrollToElementParams {
-    pub name: Option<String>,
-    pub role: Option<AxRole>,
-    pub value_contains: Option<String>,
-    /// Primary sweep direction; the search reverses to the other end if the
-    /// target isn't found first.
-    pub direction: ScrollDirection,
-    /// Scroll anchor (window-relative). `None` → the active window's center.
-    pub anchor: Option<(i32, i32)>,
-    /// Wheel notches issued per scroll step.
-    pub step: u32,
-    /// Overall wall-clock bound.
-    pub timeout_ms: u64,
-}
-
-/// Outcome of [`Glass::scroll_to_element`].
-#[derive(Clone, Debug)]
-pub struct ScrollToElementOutcome {
-    pub matched: bool,
-    /// The matched element (absent when `matched` is false). Its id is from the
-    /// final snapshot, so it is usable with `click_element`.
-    pub element: Option<ElementInfo>,
-    pub elapsed_ms: u64,
-    /// Total scroll steps issued across the sweep.
-    pub steps: u32,
-    /// Whether the sweep had reversed past the primary direction when it returned.
-    pub reversed: bool,
-}
-
-/// Parameters for [`Glass::wait_for_region`].
-#[derive(Clone, Debug)]
-pub struct WaitRegionParams {
-    /// Saved baseline to compare against; `None` uses the frame at call start.
-    pub baseline: Option<String>,
-    /// Window-relative sub-rectangle to watch; `None` watches the whole window.
-    pub region: Option<Region>,
-    pub until: RegionUntil,
-    /// `true` = perceptual diff (use `threshold`); `false` = exact (use `tolerance`).
-    pub perceptual: bool,
-    pub threshold: f32,
-    pub tolerance: u8,
-    pub interval_ms: u64,
-    pub timeout_ms: u64,
-    /// When set, watch this window's own region instead of the active window's —
-    /// without changing which window is active.
-    pub window: Option<WindowId>,
-}
-
-/// Outcome of [`Glass::wait_for_region`]. `frame` is the last captured region
-/// (window when no region), for the optional image at the MCP layer.
-#[derive(Clone, Debug)]
-pub struct WaitRegionOutcome {
-    /// Whether the region condition held before the timeout.
-    pub matched: bool,
-    /// Percent of the watched region that differed from the reference at the last poll.
-    pub changed_pct: f32,
-    /// Bounding box of the changed area at the last poll (None if nothing changed).
-    pub bbox: Option<BBox>,
-    /// The last captured region frame (the watched window when no region) — source for the optional image at the tool layer.
-    pub frame: Frame,
-    /// Wall-clock milliseconds elapsed when the wait returned.
-    pub elapsed_ms: u64,
-}
-
-/// Parameters for [`Glass::wait_for_log`].
-#[derive(Clone, Debug)]
-pub struct WaitLogParams {
-    /// Substring to wait for (required by the tool layer to be non-empty).
-    pub contains: String,
-    pub stream: Option<Stream>,
-    /// Start scanning from this cursor; `None` = the buffer's end at call start
-    /// (so only newly-appended lines count).
-    pub cursor: Option<u64>,
-    pub interval_ms: u64,
-    pub timeout_ms: u64,
-}
-
-/// Outcome of [`Glass::wait_for_log`].
-#[derive(Clone, Debug)]
-pub struct WaitLogOutcome {
-    pub matched: bool,
-    pub line: Option<LogLine>,
-    /// Cursor to resume from: just past the matched line, or the buffer end on timeout.
-    pub cursor: u64,
-    pub elapsed_ms: u64,
-    /// Set on a timeout when the substring was already in the buffer *before* this call's
-    /// start cursor (the default-cursor footgun: a fast-boot line is otherwise skipped).
-    /// Points the caller at `cursor:0` instead of failing silently.
-    pub note: Option<String>,
-}
+pub use wait::{
+    ScrollDirection, ScrollToElementOutcome, ScrollToElementParams, WaitElementOutcome,
+    WaitElementParams, WaitLogOutcome, WaitLogParams, WaitRegionOutcome, WaitRegionParams,
+    WaitStableOutcome, WaitStableParams, SCROLL_TO_DEFAULT_STEP, SCROLL_TO_DEFAULT_TIMEOUT_MS,
+};
 
 struct ActiveSession {
     platform: Box<dyn Platform + Send>,
@@ -511,896 +333,6 @@ impl Glass {
 
     fn active_mut(&mut self) -> Result<&mut ActiveSession> {
         self.active.as_mut().ok_or(GlassError::NoActiveSession)
-    }
-
-    /// Validate that any window-relative coordinates in `event` fall inside the
-    /// current window.
-    fn check_bounds(&self, event: &PointerEvent) -> Result<()> {
-        let g = self.require_active()?;
-        let (w, h) = (g.geometry.width as i32, g.geometry.height as i32);
-        let check = |x: i32, y: i32| -> Result<()> {
-            if x < 0 || y < 0 || x >= w || y >= h {
-                Err(GlassError::CoordOutOfBounds {
-                    x,
-                    y,
-                    width: g.geometry.width,
-                    height: g.geometry.height,
-                })
-            } else {
-                Ok(())
-            }
-        };
-        match *event {
-            PointerEvent::Move { x, y } => check(x, y),
-            PointerEvent::Click { x, y, .. } => check(x, y),
-            PointerEvent::Scroll { x, y, .. } => check(x, y),
-            PointerEvent::Drag {
-                from_x,
-                from_y,
-                to_x,
-                to_y,
-                ..
-            } => {
-                check(from_x, from_y)?;
-                check(to_x, to_y)
-            }
-            PointerEvent::Gesture { ref pointers, .. } => {
-                for p in pointers {
-                    check(p.from_x, p.from_y)?;
-                    check(p.to_x, p.to_y)?;
-                }
-                Ok(())
-            }
-        }
-    }
-
-    /// Start with the default backend.
-    pub fn start(&mut self, spec: &AppSpec) -> Result<WindowGeometry> {
-        let backend = self.default_backend.clone();
-        self.start_on(&backend, spec)
-    }
-
-    /// Start with an explicit backend, constructing it via the factory.
-    pub fn start_on(&mut self, backend: &str, spec: &AppSpec) -> Result<WindowGeometry> {
-        let t = std::time::Instant::now();
-        let result = self.start_on_inner(backend, spec);
-        self.emit_audit(
-            &crate::audit::Actuation::Launch { spec, backend },
-            crate::audit::AuditOutcome::from_result(&result),
-            t.elapsed(),
-        );
-        result
-    }
-
-    fn start_on_inner(&mut self, backend: &str, spec: &AppSpec) -> Result<WindowGeometry> {
-        // One active session: tear down any current one first.
-        if let Some(mut s) = self.active.take() {
-            let _ = s.platform.stop_app();
-        }
-        let Backend {
-            mut platform,
-            accessibility,
-        } = (self.factory)(backend)?;
-        let geometry = platform.start_app(spec)?;
-        let mut session = ActiveSession {
-            platform,
-            accessibility,
-            last_ax: None,
-            geometry: geometry.clone(),
-            logs: LogBuffer::new(self.log_capacity),
-            active_window: None,
-        };
-        session.pump();
-        session.active_window = session
-            .platform
-            .list_windows()
-            .ok()
-            .and_then(|ws| ws.iter().find(|w| w.active).or_else(|| ws.first()).cloned())
-            .map(|w| crate::audit::WindowRef {
-                id: w.id.0,
-                title: w.title,
-            });
-        self.active = Some(session);
-        Ok(geometry)
-    }
-
-    pub fn stop(&mut self) -> Result<()> {
-        let t = std::time::Instant::now();
-        // Snapshot the window BEFORE stop_inner, which drops self.active — so this
-        // records on the dedicated path rather than emit_audit (which would see None
-        // after teardown). Keep this ordering if refactoring, or window attribution breaks.
-        let window = self.active.as_ref().and_then(|s| s.active_window.clone());
-        let result = self.stop_inner();
-        if let Some(sink) = &self.audit {
-            sink.record(
-                &crate::audit::Actuation::Stop,
-                &crate::audit::ActuationContext { window },
-                &crate::audit::AuditOutcome::from_result(&result),
-                t.elapsed(),
-            );
-        }
-        result
-    }
-
-    fn stop_inner(&mut self) -> Result<()> {
-        let mut s = self.active.take().ok_or(GlassError::NoActiveSession)?;
-        s.platform.stop_app()
-        // `s` drops here, tearing down the spawned backend (Xvfb/sway).
-    }
-
-    /// Best-effort teardown of **all** active sessions for process exit. Idempotent:
-    /// a no-op when nothing is active. Errors are swallowed — we are exiting, so a
-    /// failed `stop_app` must not prevent releasing the rest (the OS reaps anything
-    /// left). Distinct from `stop()`, which reports errors to a tool caller.
-    ///
-    /// Written to drain the session set so the future multi-session registry (a
-    /// `HashMap` instead of this `Option`) reuses it unchanged — it becomes a `for`
-    /// loop with no other change.
-    pub fn shutdown(&mut self) {
-        if let Some(mut s) = self.active.take() {
-            let _ = s.platform.stop_app();
-            // `s` drops here: the backend (Xvfb/sway/Job) is torn down.
-        }
-        if let Some(hook) = self.shutdown_hook.take() {
-            hook();
-        }
-    }
-
-    pub fn geometry(&self) -> Result<WindowGeometry> {
-        Ok(self.require_active()?.geometry.clone())
-    }
-
-    /// Capture the active window, or — when `window` is set — a different
-    /// window's region WITHOUT changing which window is active (unlike
-    /// `select_window`). `region` is relative to whichever window is captured.
-    pub fn screenshot(
-        &mut self,
-        region: Option<Region>,
-        window: Option<WindowId>,
-    ) -> Result<Frame> {
-        self.capture(window, region.as_ref())
-    }
-
-    /// Capture `window`'s region (or, when `None`, the active window's), pumping
-    /// logs afterward either way. A specific window's own geometry governs its
-    /// capture — the backend validates `id` and any region against it — so the
-    /// active window's cached `s.geometry` is only consulted for the `None` case.
-    fn capture(&mut self, window: Option<WindowId>, region: Option<&Region>) -> Result<Frame> {
-        let s = self.active_mut()?;
-        let frame = match window {
-            Some(id) => s.platform.capture_window(id, region)?,
-            None => {
-                if let Some(r) = region {
-                    r.check_fits(s.geometry.width, s.geometry.height)?;
-                }
-                s.platform.capture_frame(region)?
-            }
-        };
-        s.pump();
-        Ok(frame)
-    }
-
-    pub fn pointer(&mut self, event: &PointerEvent) -> Result<()> {
-        let t = std::time::Instant::now();
-        let result = self.pointer_inner(event);
-        self.emit_audit(
-            &crate::audit::Actuation::Pointer { event },
-            crate::audit::AuditOutcome::from_result(&result),
-            t.elapsed(),
-        );
-        result
-    }
-
-    fn pointer_inner(&mut self, event: &PointerEvent) -> Result<()> {
-        self.check_bounds(event)?;
-        let s = self.active_mut()?;
-        s.platform.send_pointer(event)?;
-        s.pump();
-        Ok(())
-    }
-
-    pub fn key(&mut self, event: &KeyEvent) -> Result<()> {
-        let t = std::time::Instant::now();
-        let result = self.key_inner(event);
-        self.emit_audit(
-            &crate::audit::Actuation::Key { event },
-            crate::audit::AuditOutcome::from_result(&result),
-            t.elapsed(),
-        );
-        result
-    }
-
-    fn key_inner(&mut self, event: &KeyEvent) -> Result<()> {
-        let s = self.active_mut()?;
-        s.platform.send_key(event)?;
-        s.pump();
-        Ok(())
-    }
-
-    pub fn get_clipboard(&mut self) -> Result<String> {
-        self.active_mut()?.platform.get_clipboard()
-    }
-
-    pub fn set_clipboard(&mut self, text: &str) -> Result<()> {
-        let t = std::time::Instant::now();
-        let result = self.set_clipboard_inner(text);
-        self.emit_audit(
-            &crate::audit::Actuation::ClipboardSet { text },
-            crate::audit::AuditOutcome::from_result(&result),
-            t.elapsed(),
-        );
-        result
-    }
-
-    fn set_clipboard_inner(&mut self, text: &str) -> Result<()> {
-        self.active_mut()?.platform.set_clipboard(text)
-    }
-
-    pub fn window(&mut self, op: &WindowOp) -> Result<WindowGeometry> {
-        let t = std::time::Instant::now();
-        let result = self.window_inner(op);
-        if !matches!(op, WindowOp::Geometry) {
-            self.emit_audit(
-                &crate::audit::Actuation::Window { op },
-                crate::audit::AuditOutcome::from_result(&result),
-                t.elapsed(),
-            );
-        }
-        result
-    }
-
-    fn window_inner(&mut self, op: &WindowOp) -> Result<WindowGeometry> {
-        let s = self.active_mut()?;
-        let geometry = s.platform.window(op)?;
-        s.geometry = geometry.clone();
-        s.pump();
-        Ok(geometry)
-    }
-
-    /// All top-level windows of the active app.
-    pub fn list_windows(&mut self) -> Result<Vec<WindowInfo>> {
-        self.active_mut()?.platform.list_windows()
-    }
-
-    /// Make `id` the active window; subsequent capture/input/window ops target
-    /// it. Updates the cached active-window geometry.
-    pub fn select_window(&mut self, id: WindowId) -> Result<WindowGeometry> {
-        let s = self.active_mut()?;
-        let geometry = s.platform.select_window(id)?;
-        s.geometry = geometry.clone();
-        s.active_window = Some(crate::audit::WindowRef {
-            id: id.0,
-            title: None,
-        });
-        s.pump();
-        Ok(geometry)
-    }
-
-    /// Snapshot the active window's accessibility tree (normalized, window-
-    /// relative, ids assigned by the core). Caches it for `click_element`.
-    /// `AxUnsupported` if the backend has no accessibility reader.
-    pub fn a11y_snapshot(&mut self) -> Result<AxTree> {
-        let s = self.active_mut()?;
-        let pids = s.platform.app_pids();
-        let window = s.geometry.clone();
-        let window_handle = s.platform.active_window_handle();
-        let a11y_bus_addr = s.platform.a11y_bus_addr();
-        let acc = s.accessibility.as_mut().ok_or(GlassError::AxUnsupported)?;
-        let mut tree = acc.snapshot(&AxContext {
-            pids,
-            window,
-            window_handle,
-            a11y_bus_addr,
-        })?;
-        tree.assign_ids();
-        s.last_ax = Some(tree.clone());
-        s.pump();
-        Ok(tree)
-    }
-
-    /// Capture the active window and overlay numbered marks on its interactable
-    /// accessibility elements. Returns the annotated frame and the marks legend.
-    /// Caches the snapshot, so `click_element` resolves a mark's id afterward.
-    pub fn a11y_marks(&mut self) -> Result<(Frame, Vec<Mark>)> {
-        let frame = self.screenshot(None, None)?;
-        let tree = self.a11y_snapshot()?;
-        Ok(crate::marks::render(&frame, &tree))
-    }
-
-    /// Click the element with id `id` from the most recent `a11y_snapshot`
-    /// (clicks the center of its bounds, via the normal pointer path).
-    pub fn click_element(&mut self, id: AxNodeId) -> Result<()> {
-        let t = std::time::Instant::now();
-        let element = self.element_ref(id);
-        let result = self.click_element_inner(id);
-        self.emit_audit(
-            &crate::audit::Actuation::ClickElement { element },
-            crate::audit::AuditOutcome::from_result(&result),
-            t.elapsed(),
-        );
-        result
-    }
-
-    fn click_element_inner(&mut self, id: AxNodeId) -> Result<()> {
-        let (bounds, active_geo) = {
-            let s = self.require_active()?;
-            let tree = s.last_ax.as_ref().ok_or(GlassError::NoAxSnapshot)?;
-            let node = tree.find(id).ok_or(GlassError::AxElementNotFound(id.0))?;
-            let bounds = node.bounds.ok_or(GlassError::AxElementNotClickable(id.0))?;
-            (bounds, s.geometry.clone())
-        };
-        // The element's a11y bounds are reported relative to the active window, but it
-        // may actually render in a separate popover window (e.g. an open dropdown's
-        // option list) whose own origin they don't reflect. Detect that and route the
-        // click into the popover instead of silently missing.
-        //
-        // This enumeration is a best-effort popover probe, not something an ordinary
-        // click depends on: a backend where `list_windows` is heavier or flaky must
-        // never turn a normal click into a failure just because the probe failed. An
-        // `Err` here degrades to an empty list, which makes `owning_popover` return
-        // `None` below and falls straight through to the unchanged `clamped_center`
-        // click path.
-        let windows = self.list_windows().unwrap_or_default();
-        if let Some(popover_id) = owning_popover(bounds, &active_geo, &windows) {
-            let popover_geo = windows
-                .iter()
-                .find(|w| w.id == popover_id)
-                .map(|w| w.geometry.clone())
-                .ok_or(GlassError::WindowNotFound)?;
-            let container = {
-                let s = self.require_active()?;
-                let tree = s.last_ax.as_ref().ok_or(GlassError::NoAxSnapshot)?;
-                menu_container_bounds(&tree.root, id, &popover_geo)
-            }
-            .ok_or(GlassError::AxElementInUnmappedPopover(id.0))?;
-            let prev = windows.iter().find(|w| w.active).map(|w| w.id);
-            self.select_window(popover_id)?;
-            let result = self.pointer_inner(&PointerEvent::Click {
-                x: bounds.x - container.x,
-                y: bounds.y - container.y,
-                button: MouseButton::Left,
-                count: 1,
-                modifiers: vec![],
-            });
-            // Best-effort restore: the click's own result (ok or err) still wins.
-            if let Some(prev) = prev {
-                let _ = self.select_window(prev);
-            }
-            return result;
-        }
-        let (x, y) = bounds
-            .clamped_center(active_geo.width, active_geo.height)
-            .ok_or(GlassError::AxElementNotClickable(id.0))?;
-        self.pointer_inner(&PointerEvent::Click {
-            x,
-            y,
-            button: MouseButton::Left,
-            count: 1,
-            modifiers: vec![],
-        })
-    }
-
-    /// Set the value/text of element `id` (from the latest `a11y_snapshot`) via the
-    /// platform a11y API. Errors `NoAxSnapshot`/`AxElementNotFound` (id not in the
-    /// cached snapshot), `AxUnsupported` (no reader), or — from the backend —
-    /// `AxElementNotEditable`/`AxElementChanged`.
-    pub fn set_value(&mut self, id: AxNodeId, text: &str) -> Result<()> {
-        let t = std::time::Instant::now();
-        let element = self.element_ref(id);
-        let result = self.set_value_inner(id, text);
-        self.emit_audit(
-            &crate::audit::Actuation::SetValue { element, text },
-            crate::audit::AuditOutcome::from_result(&result),
-            t.elapsed(),
-        );
-        result
-    }
-
-    fn set_value_inner(&mut self, id: AxNodeId, text: &str) -> Result<()> {
-        let (target, ctx) = {
-            let s = self.require_active()?;
-            // Check for reader availability before consulting the snapshot, so that
-            // `AxUnsupported` takes precedence when there is no accessibility backend.
-            if s.accessibility.is_none() {
-                return Err(GlassError::AxUnsupported);
-            }
-            let tree = s.last_ax.as_ref().ok_or(GlassError::NoAxSnapshot)?;
-            let node = tree.find(id).ok_or(GlassError::AxElementNotFound(id.0))?;
-            let target = AxTarget {
-                id,
-                role: node.role,
-                name: node.name.clone(),
-                bounds: node.bounds,
-            };
-            let ctx = AxContext {
-                pids: s.platform.app_pids(),
-                window: s.geometry.clone(),
-                window_handle: s.platform.active_window_handle(),
-                a11y_bus_addr: s.platform.a11y_bus_addr(),
-            };
-            (target, ctx)
-        };
-        // A dropdown/combo has no committing accessibility write: its `Selection`
-        // interface only moves the popup's *preview* selection, and the model commits
-        // only on row activation (Enter/click). So drive it like a person does —
-        // open it, keyboard-navigate to the option, and press Enter.
-        if target.role == AxRole::ComboBox {
-            return self.set_combo_value(id, &target, text);
-        }
-        let s = self.active_mut()?;
-        s.accessibility
-            .as_mut()
-            .ok_or(GlassError::AxUnsupported)?
-            .set_value(&ctx, &target, text)?;
-        s.pump();
-        Ok(())
-    }
-
-    /// Select an option in a dropdown/combo by label (case-insensitive). Opens the
-    /// popup, arrow-navigates from the current selection to the target, and presses
-    /// Enter to commit — verifying the button label changed (else `AxValueNotApplied`).
-    fn set_combo_value(&mut self, id: AxNodeId, target: &AxTarget, text: &str) -> Result<()> {
-        let want = text.trim();
-        // Already showing it? (the combo's name is its current selection label)
-        if target
-            .name
-            .as_deref()
-            .is_some_and(|n| n.eq_ignore_ascii_case(want))
-        {
-            return Ok(());
-        }
-        // Open the popup (the combo button is in the main window, so this click lands).
-        self.click_element(id)?;
-        self.settle_for_popup();
-        // Re-read the realized options + which one is currently selected. The open
-        // combo is `expanded`; when several combos exist, ids don't survive the
-        // re-snapshot, so fall back to the one nearest the target's bounds.
-        let tree = self.a11y_snapshot()?;
-        let combo = find_expanded_combo(&tree.root)
-            .or_else(|| find_combo_near(&tree.root, target.bounds.as_ref()))
-            .ok_or(GlassError::AxElementChanged(id.0))?;
-        let options = collect_combo_options(combo);
-        if options.is_empty() {
-            return Err(GlassError::AxElementNotEditable(id.0));
-        }
-        let target_idx = options
-            .iter()
-            .position(|(label, _)| label.eq_ignore_ascii_case(want));
-        let Some(target_idx) = target_idx else {
-            // Unknown option — dismiss the popup so the UI is left neutral, then report.
-            let _ = self.key(&KeyEvent::Chord("Escape".to_string()));
-            let choices = options
-                .iter()
-                .map(|(l, _)| l.clone())
-                .collect::<Vec<_>>()
-                .join(", ");
-            return Err(GlassError::AxOptionNotFound(
-                id.0,
-                text.to_string(),
-                choices,
-            ));
-        };
-        // Opening focuses the current selection; step from it to the target, then Enter.
-        let current_idx = options.iter().position(|(_, sel)| *sel).unwrap_or(0);
-        let delta = target_idx as i32 - current_idx as i32;
-        let chord = if delta >= 0 { "Down" } else { "Up" };
-        for _ in 0..delta.unsigned_abs() {
-            self.key(&KeyEvent::Chord(chord.to_string()))?;
-        }
-        self.key(&KeyEvent::Chord("Return".to_string()))?;
-        self.settle_for_popup();
-        // Verify the model actually committed — the *target* combo (matched by bounds,
-        // now closed so nothing is `expanded`) must read the wanted label.
-        let tree = self.a11y_snapshot()?;
-        let ok = find_combo_near(&tree.root, target.bounds.as_ref())
-            .and_then(|c| c.name.as_deref())
-            .is_some_and(|n| n.eq_ignore_ascii_case(want));
-        if ok {
-            Ok(())
-        } else {
-            Err(GlassError::AxValueNotApplied(id.0))
-        }
-    }
-
-    /// Let a just-opened/closed popup realize in the a11y tree before re-reading.
-    fn settle_for_popup(&self) {
-        std::thread::sleep(std::time::Duration::from_millis(250));
-    }
-
-    pub fn wait_stable(&mut self, params: &WaitStableParams) -> Result<WaitStableOutcome> {
-        let active = self.require_active()?;
-        // The active window's cached geometry only bounds a stability_region when
-        // watching the active window itself; a specific `window` is validated by
-        // the backend against its own geometry instead (see `capture`).
-        if params.window.is_none() {
-            let geo = active.geometry.clone();
-            if let Some(r) = &params.stability_region {
-                r.check_fits(geo.width, geo.height)?;
-            }
-        }
-        let mut tracker = StabilityTracker::new(params.settle_frames, params.tolerance);
-        let region = params.stability_region;
-        let window = params.window;
-        let outcome = crate::poll::poll_until(params.interval_ms, params.timeout_ms, || {
-            // Poll only the watched region (cheap) when one is set; else the full window.
-            let frame = self.capture(window, region.as_ref())?;
-            let settled = tracker.observe(frame)?;
-            Ok(if settled { Some(()) } else { None })
-        })?;
-        let settled = outcome.value.is_some();
-        // Return the full window: a fresh capture if we were polling a sub-region
-        // (the genuinely-settled state), else the just-observed full frame.
-        let frame = match region {
-            Some(_) => self.capture(window, None)?,
-            None => tracker.last().cloned().expect("a frame was just observed"),
-        };
-        Ok(WaitStableOutcome {
-            frame,
-            settled,
-            saw_motion: tracker.saw_change(),
-            observed_ms: outcome.elapsed_ms,
-        })
-    }
-
-    /// Block until a precise accessibility-element condition holds, re-snapshotting
-    /// each tick. Text-only outcome. The final snapshot is cached (so the returned
-    /// element id is immediately usable with `click_element`). Errors immediately if
-    /// the backend has no accessibility reader (the first snapshot fails).
-    pub fn wait_for_element(&mut self, params: &WaitElementParams) -> Result<WaitElementOutcome> {
-        self.require_active()?; // fail fast; a11y_snapshot rechecks inside the loop
-        let outcome = crate::poll::poll_until(params.interval_ms, params.timeout_ms, || {
-            let tree = self.a11y_snapshot()?; // fresh snapshot; assigns ids, caches, pumps
-            Ok(
-                match element_match(
-                    &tree,
-                    params.name.as_deref(),
-                    params.role,
-                    params.value_contains.as_deref(),
-                    params.condition,
-                ) {
-                    ElementMatch::Satisfied(node) => Some(node.map(ElementInfo::from_node)),
-                    ElementMatch::Pending => None,
-                },
-            )
-        })?;
-        Ok(WaitElementOutcome {
-            matched: outcome.value.is_some(),
-            element: outcome.value.flatten(),
-            elapsed_ms: outcome.elapsed_ms,
-        })
-    }
-
-    /// Scroll a container (at `anchor`, default the active window's center) until an
-    /// element matching name/role/value realizes in the a11y tree, then return it —
-    /// its id is from the final snapshot, so it is immediately `click_element`-able.
-    /// For a virtualized list the target row is absent from the tree until scrolled
-    /// into range; this checks the current view, sweeps the primary `direction` to
-    /// its end, then reverses to cover the other end. End-of-scroll is detected from
-    /// the accessibility tree: when a scroll step leaves the tree's outline unchanged,
-    /// the container did not advance (immune to cosmetic repaints — a scroller's
-    /// boundary shadow, a focus ring, a blinking caret — that a pixel-motion signal
-    /// would misread as "still scrolling"). A target never realized after a full
-    /// bidirectional sweep or `timeout_ms` yields a soft `{matched:false}` (not an
-    /// error), like `wait_for_element`. The scroll actions are audited via the pointer
-    /// path; there is no separate top-level audit entry.
-    ///
-    /// Limitations of the a11y-tree end-of-scroll signal: (1) a container holding a
-    /// continuously-repainting a11y node — a live region, a clock, a progress bar —
-    /// never leaves the tree "unchanged", so the sweep runs to `timeout_ms` in the
-    /// primary direction and returns `{matched:false}` instead of reversing; pass the
-    /// `direction` the target actually lies in to avoid the wasted sweep. (2) A very
-    /// long list can exceed `timeout_ms` before a distant target scrolls into range —
-    /// raise `timeout_ms`, or `step` to cover more per move.
-    pub fn scroll_to_element(
-        &mut self,
-        params: &ScrollToElementParams,
-    ) -> Result<ScrollToElementOutcome> {
-        self.require_active()?;
-        let start = std::time::Instant::now();
-        let geo = self.geometry()?;
-        let (ax, ay) = params
-            .anchor
-            .unwrap_or((geo.width as i32 / 2, geo.height as i32 / 2));
-
-        // Snapshot the current view: return immediately if already realized, and seed
-        // the outline the first scroll step is compared against.
-        let (found, mut prev_outline) = self.snapshot_match_outline(params)?;
-        if let Some(info) = found {
-            return Ok(ScrollToElementOutcome {
-                matched: true,
-                element: Some(info),
-                elapsed_ms: start.elapsed().as_millis() as u64,
-                steps: 0,
-                reversed: false,
-            });
-        }
-
-        let mut steps: u32 = 0;
-        for (i, dir) in [params.direction, params.direction.opposite()]
-            .into_iter()
-            .enumerate()
-        {
-            let reversed = i == 1;
-            loop {
-                if start.elapsed().as_millis() as u64 >= params.timeout_ms
-                    || steps >= SCROLL_TO_MAX_STEPS
-                {
-                    return Ok(ScrollToElementOutcome {
-                        matched: false,
-                        element: None,
-                        elapsed_ms: start.elapsed().as_millis() as u64,
-                        steps,
-                        reversed,
-                    });
-                }
-                self.pointer(&PointerEvent::Scroll {
-                    x: ax,
-                    y: ay,
-                    dx: 0,
-                    dy: dir.dy(params.step),
-                    modifiers: vec![],
-                })?;
-                steps += 1;
-                // Let the scrolled rows realize in the a11y tree before re-reading.
-                std::thread::sleep(std::time::Duration::from_millis(SCROLL_TO_SETTLE_MS));
-                let (found, outline) = self.snapshot_match_outline(params)?;
-                if let Some(info) = found {
-                    return Ok(ScrollToElementOutcome {
-                        matched: true,
-                        element: Some(info),
-                        elapsed_ms: start.elapsed().as_millis() as u64,
-                        steps,
-                        reversed,
-                    });
-                }
-                // No change in the a11y tree ⇒ the container did not advance ⇒ this
-                // end is reached; sweep the opposite direction.
-                let saturated = outline == prev_outline;
-                prev_outline = outline;
-                if saturated {
-                    break;
-                }
-            }
-        }
-        Ok(ScrollToElementOutcome {
-            matched: false,
-            element: None,
-            elapsed_ms: start.elapsed().as_millis() as u64,
-            steps,
-            reversed: true,
-        })
-    }
-
-    /// Snapshot the current view once; return the matched element (if the selector is
-    /// satisfied) and the tree's outline. The snapshot is cached, so a returned
-    /// element's id is usable with `click_element`. The outline is the end-of-scroll
-    /// signal: unchanged across a scroll step ⇒ the container did not advance.
-    fn snapshot_match_outline(
-        &mut self,
-        params: &ScrollToElementParams,
-    ) -> Result<(Option<ElementInfo>, String)> {
-        let tree = self.a11y_snapshot()?;
-        let found = match element_match(
-            &tree,
-            params.name.as_deref(),
-            params.role,
-            params.value_contains.as_deref(),
-            ElementCondition::Appears,
-        ) {
-            ElementMatch::Satisfied(node) => node.map(ElementInfo::from_node),
-            ElementMatch::Pending => None,
-        };
-        Ok((found, tree.to_outline()))
-    }
-
-    /// Block until a watched region diverges from / converges to a reference.
-    /// Compares in-memory each tick (no WebP encode). Text-only outcome; the last
-    /// captured frame is returned for an optional image at the tool layer.
-    /// If `baseline` is set and `region` is `None`, the baseline must match the
-    /// current window size — a size change since it was saved returns `SizeMismatch`;
-    /// crop to a stable `region` to avoid this.
-    pub fn wait_for_region(&mut self, params: &WaitRegionParams) -> Result<WaitRegionOutcome> {
-        let active = self.require_active()?;
-        // As in `wait_stable`: the active window's cached geometry only bounds
-        // `region` when watching the active window; a specific `window` is
-        // validated by the backend against its own geometry instead.
-        if params.window.is_none() {
-            let geo = active.geometry.clone();
-            if let Some(r) = &params.region {
-                r.check_fits(geo.width, geo.height)?;
-            }
-        }
-        // Reference: a saved baseline (cropped to the region) or the current frame.
-        let reference: Frame = match &params.baseline {
-            Some(name) => {
-                let base = self.baselines.load(name)?;
-                match &params.region {
-                    Some(r) => base.crop(r)?,
-                    None => base,
-                }
-            }
-            None => self.capture(params.window, params.region.as_ref())?,
-        };
-        let (perceptual, threshold, tolerance, until, region, window) = (
-            params.perceptual,
-            params.threshold,
-            params.tolerance,
-            params.until,
-            params.region,
-            params.window,
-        );
-        let mut last: Option<(f32, Option<BBox>, Frame)> = None;
-        let outcome = crate::poll::poll_until(params.interval_ms, params.timeout_ms, || {
-            let current = self.capture(window, region.as_ref())?;
-            let d = if perceptual {
-                diff_perceptual(&reference, &current, threshold)?
-            } else {
-                diff(&reference, &current, tolerance)?
-            };
-            let satisfied = region_satisfied(&d, until);
-            last = Some((d.changed_pct, d.bbox, current));
-            Ok(if satisfied { Some(()) } else { None })
-        })?;
-        let (changed_pct, bbox, frame) = last.expect("at least one poll ran");
-        Ok(WaitRegionOutcome {
-            matched: outcome.value.is_some(),
-            changed_pct,
-            bbox,
-            frame,
-            elapsed_ms: outcome.elapsed_ms,
-        })
-    }
-
-    /// Block until a log line matching `contains` (and optional stream) appears,
-    /// scanning from `cursor` (default: the buffer end at call start, so only new
-    /// lines count). Returns the matched line and a resume cursor; on timeout
-    /// returns `{matched:false}` with the current end cursor.
-    pub fn wait_for_log(&mut self, params: &WaitLogParams) -> Result<WaitLogOutcome> {
-        let start_cursor = {
-            let s = self.active_mut()?;
-            s.pump();
-            params.cursor.unwrap_or_else(|| s.logs.end_cursor())
-        };
-        let (contains, stream) = (params.contains.clone(), params.stream);
-        let mut scan_cursor = start_cursor;
-        let outcome = crate::poll::poll_until(params.interval_ms, params.timeout_ms, || {
-            let s = self.active_mut()?;
-            s.pump();
-            let (lines, next) = s.logs.read(scan_cursor, 1, stream, Some(&contains));
-            scan_cursor = next; // advance past already-examined lines so we don't re-scan
-            Ok(lines.into_iter().next())
-        })?;
-        let s = self.active_mut()?;
-        s.pump();
-        let end = s.logs.end_cursor();
-        Ok(match outcome.value {
-            Some(line) => WaitLogOutcome {
-                cursor: line.seq + 1,
-                line: Some(line),
-                matched: true,
-                elapsed_ms: outcome.elapsed_ms,
-                note: None,
-            },
-            None => {
-                // The default cursor is the buffer end at call start, so a line emitted
-                // *before* this call (e.g. a fast-boot "ready") is skipped and we time out.
-                // If the substring is already in the buffer before our start cursor, say so
-                // rather than failing silently — point the caller at cursor:0.
-                let note = if params.cursor.is_none() {
-                    let (earlier, _) = s.logs.read(0, 1, stream, Some(&contains));
-                    earlier
-                        .into_iter()
-                        .next()
-                        .filter(|l| l.seq < start_cursor)
-                        .map(|l| {
-                            format!(
-                                "{contains:?} was already in the log at seq {} (before this call); \
-                                 pass cursor:0 to match already-buffered lines",
-                                l.seq
-                            )
-                        })
-                } else {
-                    None
-                };
-                WaitLogOutcome {
-                    matched: false,
-                    line: None,
-                    cursor: end,
-                    elapsed_ms: outcome.elapsed_ms,
-                    note,
-                }
-            }
-        })
-    }
-
-    pub fn save_baseline(&mut self, name: &str) -> Result<()> {
-        let frame = {
-            let s = self.active_mut()?;
-            let frame = s.platform.capture_frame(None)?;
-            s.pump();
-            frame
-        };
-        self.baselines.save(name, &frame)
-    }
-
-    /// Load the named baseline and capture the current window frame, both scoped
-    /// to `region` when set (the whole window otherwise). Baselines are stored
-    /// whole and cropped here, so one saved baseline can be compared against any
-    /// sub-region — and both operands are always cropped consistently, never
-    /// silently mismatched.
-    fn baseline_and_current(
-        &mut self,
-        name: &str,
-        region: Option<&Region>,
-    ) -> Result<(Frame, Frame)> {
-        if let Some(r) = region {
-            let geo = self.require_active()?.geometry.clone();
-            r.check_fits(geo.width, geo.height)?;
-        }
-        let baseline = {
-            let base = self.baselines.load(name)?;
-            match region {
-                Some(r) => base.crop(r)?,
-                None => base,
-            }
-        };
-        let current = {
-            let s = self.active_mut()?;
-            let frame = s.platform.capture_frame(region)?;
-            s.pump();
-            frame
-        };
-        Ok((baseline, current))
-    }
-
-    /// Exact per-channel diff of the current frame against a saved baseline.
-    /// `region` scopes the comparison to a window-relative sub-rectangle.
-    pub fn diff_baseline(
-        &mut self,
-        name: &str,
-        region: Option<&Region>,
-        tolerance: u8,
-    ) -> Result<DiffResult> {
-        self.diff_baseline_with_frame(name, region, tolerance)
-            .map(|(r, _)| r)
-    }
-
-    /// Like [`diff_baseline`] but also returns the current frame that was compared.
-    pub fn diff_baseline_with_frame(
-        &mut self,
-        name: &str,
-        region: Option<&Region>,
-        tolerance: u8,
-    ) -> Result<(DiffResult, Frame)> {
-        let (baseline, current) = self.baseline_and_current(name, region)?;
-        let r = diff(&baseline, &current, tolerance)?;
-        Ok((r, current))
-    }
-
-    /// Perceptual diff (YIQ + anti-alias suppression) against a saved baseline —
-    /// the default for regression, robust to anti-aliasing / sub-pixel / GPU-font
-    /// rendering noise. `threshold` ∈ [0,1] (smaller = stricter). `region` scopes
-    /// the comparison to a window-relative sub-rectangle.
-    pub fn diff_baseline_perceptual(
-        &mut self,
-        name: &str,
-        region: Option<&Region>,
-        threshold: f32,
-    ) -> Result<DiffResult> {
-        self.diff_baseline_perceptual_with_frame(name, region, threshold)
-            .map(|(r, _)| r)
-    }
-
-    /// Like [`diff_baseline_perceptual`] but also returns the current frame compared.
-    pub fn diff_baseline_perceptual_with_frame(
-        &mut self,
-        name: &str,
-        region: Option<&Region>,
-        threshold: f32,
-    ) -> Result<(DiffResult, Frame)> {
-        let (baseline, current) = self.baseline_and_current(name, region)?;
-        let r = diff_perceptual(&baseline, &current, threshold)?;
-        Ok((r, current))
     }
 
     pub fn logs(
